+++ src/test/java/subway/RestAssuredTest.java
@@ -10,7 +10,6 @@ import org.springframework.http.HttpStatus;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 public class RestAssuredTest {
 
     @DisplayName("구글 페이지 접근 테스트")
